FEAT: 구독한 언론사 탭에서의 swipe 구현
