+++ components/news/subscribeNews/subscribeNews.js
@@ -1,8 +1,8 @@
-import { allNewsData } from "../../data/allNewsData.js";
-import { shortenStr } from "../../util/shortenStr.js";
+import { subscribeData } from "../../../data/subscribeData.js";
+import { shortenStr } from "../../../util/shortenStr.js";
 
-export const allNews = (cateId, brandId) => {
-  const newsData = allNewsData[cateId].data[brandId];
+export const subscribeNews = (brandId) => {
+  const newsData = subscribeData[brandId];
 
   // press news selector
   const pressNews = document.querySelector(".press-news");
@@ -18,11 +18,8 @@ export const allNews = (cateId, brandId) => {
   // sub news selector
   const subNewsList = pressNews.querySelector(".sub-news-list");
 
-  // brand-id 저장
+  // news Id 저장
   pressNews.dataset.brandId = brandId;
-
-  // cate-id 저장
-  pressNews.dataset.cateId = cateId;
 
   // 언론사 로고
   brandLogo.src = newsData.brandImg;
