+++ components/swipe/allSwipe.js
@@ -1,0 +1,4 @@
+import { subscribeData } from "../../data/subscribeData.js";
+import { subscribeNews } from "../news/subscribeNews/subscribeNews.js";
+
+export const allSwipe = () => {};
