+++ index.js
@@ -1,7 +1,5 @@
 import { subscribeSwitchBtn } from "./components/subscribeSwitchBtn/subscribeSwitchBtn.js";
 import { header } from "./components/header/header.js";
-// import { swipe } from "./components/swipe/swipe.js";
 
 header();
 subscribeSwitchBtn();
-// swipe();
